Fully updated Example and minor lib changes
Example is now updated to show how to use the Lib since the changes to
the Nessus API.

README has been updated to show the new code, along with what is
supported.

DownlaodScan now returns a byte array instead of a string. There was no
reason for the cast so it was removed.

ScanStatus now returns the status in string format. Previously nothing
was returned and the status was just printed to the screen.